Bugfixing de Animaciones en UI y Texto
