+++ Rts/Assets/Animations/UI/Pausa/MenuCerrarPausa.anim
@@ -248,7 +248,7 @@ AnimationClip:
   m_HasMotionFloatCurves: 0
   m_Events:
   - time: 0.6666667
-    functionName: TriggerEvent
+    functionName: AnimationDoDone
     data: 
     objectReferenceParameter: {fileID: 0}
     floatParameter: 0
