+++ Rts/Assets/Script/Pause/TextAnimation.cs
@@ -1,0 +1,44 @@
+using System.Collections;
+using System.Collections.Generic;
+using TMPro;
+using UnityEngine;
+
+public class TextAnimation : MonoBehaviour
+{
+    TMP_Text textMesh;
+
+    Mesh mesh;
+
+    Vector3[] vertices;
+
+    [SerializeField] private float _intensity = 3.3f;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        textMesh = GetComponent<TMP_Text>();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        textMesh.ForceMeshUpdate();
+        mesh = textMesh.mesh;
+        vertices = mesh.vertices;
+
+        for (int i = 0; i < vertices.Length; i++)
+        {
+            Vector3 offset = Wobble(Time.time + i);
+
+            vertices[i] = vertices[i] + offset;
+        }
+
+        mesh.vertices = vertices;
+        textMesh.canvasRenderer.SetMesh(mesh);
+    }
+
+    Vector2 Wobble(float time)
+    {
+        return new Vector2(Mathf.Sin(time * _intensity), Mathf.Cos(time * 2.5f));
+    }
+}
