+++ Rts/Assets/Script/Pause/TextAnimation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b61d8c2f9ccb2ed4cad714759769f166
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
